+++ R/qsip_objects.R
@@ -580,3 +580,41 @@ S7::method(get_dataframe, qsip_data) <- function(x, type, original_headers = FAL
   # print(d)
   get_dataframe(d, original_headers = original_headers)
 }
+
+
+
+
+# extending print methods
+
+S7::method(print, qsip_source_data) <- function(x, ...) {
+  sd = x@data
+  print(glue::glue_col("<qSIP2::qsip_source_data>
+                       source_material_id count: {green {length(unique(sd$source_mat_id))}}"))
+}
+
+S7::method(print, qsip_sample_data) <- function(x, ...) {
+  sd = x@data
+  print(glue::glue_col("<qSIP2::qsip_sample_data>
+                       source_material_id count: {green {length(unique(sd$source_mat_id))}}
+                       sample_id count: {green {length(unique(sd$sample_id))}}"))
+}
+
+S7::method(print, qsip_feature_data) <- function(x, ...) {
+  print(glue::glue_col("<qSIP2::qsip_feature_data>
+                       feature_id count: {green {dim(x@data)[1]}}
+                       sample_id count: {green {dim(x@data)[2] - 1}}
+                       data type: {green {x@type}}"))
+}
+
+S7::method(print, qsip_data) <- function(x, ...) {
+
+
+  print(glue::glue_col("<qSIP2::qsip_data>
+                       group: {green {ifelse(is.null(x@filter_results$group), 'none', x@filter_results$group)}}
+                       feature_id count: {green {length(get_feature_ids(x, filtered = is_qsip_filtered(x)))} of {dim(x@feature_data@data)[1]}}
+                       sample_id count: {green {length(unique(x@sample_data@data$sample_id))}}
+                       filtered: {green {is_qsip_filtered(x)}}
+                       resampled: {green {is_qsip_resampled(x)}}
+                       EAF: {green {is_qsip_EAF(x)}}
+                       growth: {green {is_qsip_growth(x)}}"))
+}